+++ src/backend/src/signer.rs
@@ -12,6 +12,7 @@ use ic_cdk::api::management_canister::{
 use ic_cycles_ledger_client::{Account, ApproveArgs, ApproveError, Service as CyclesLedgerService};
 use ic_ledger_types::Subaccount;
 use serde_bytes::ByteBuf;
+use shared::types::signer::topup::{TopUpCyclesLedgerRequest, TopUpCyclesLedgerResult};
 
 #[derive(CandidType, Deserialize, Debug, Clone, Eq, PartialEq)]
 pub enum AllowSigningError {
@@ -140,52 +141,10 @@ pub async fn btc_principal_to_p2wpkh_address(
     }
 }
 
-/// A request to top up the cycles ledger.
-#[derive(CandidType, Deserialize, Debug, Clone, Eq, PartialEq, Default)]
-pub struct TopUpCyclesLedgerRequest {
-    /// If the cycles ledger account balance is below this threshold, top it up.
-    pub threshold: Option<Nat>,
-    /// The percentage of the backend canister's own cycles to send to the cycles ledger.
-    pub percentage: Option<u8>,
-}
-impl TopUpCyclesLedgerRequest {
-    pub fn threshold(&self) -> Nat {
-        self.threshold
-            .clone()
-            .unwrap_or(Nat::from(DEFAULT_CYCLES_LEDGER_TOP_UP_THRESHOLD))
-    }
-    pub fn percentage(&self) -> u8 {
-        self.percentage
-            .unwrap_or(DEFAULT_CYCLES_LEDGER_TOP_UP_PERCENTAGE)
-    }
-}
-/// The default cycles ledger top up threshold.  If the cycles ledger balance falls below this, it should be topped up.
-pub const DEFAULT_CYCLES_LEDGER_TOP_UP_THRESHOLD: u128 = 10_000_000_000_000; // 10T
-/// The proportion of the backend canitster's own cycles to send to the cycles ledger.
-pub const DEFAULT_CYCLES_LEDGER_TOP_UP_PERCENTAGE: u8 = 50;
-
-/// Possible error conditions when topping up the cycles ledger.
-#[derive(CandidType, Deserialize, Debug, Clone, Eq, PartialEq)]
-pub enum TopUpCyclesLedgerError {
-    CouldNotGetBalanceFromCyclesLedger,
-    CouldNotTopUpCyclesLedger { available: Nat, tried_to_send: Nat },
-}
-/// Possible successful responses when topping up the cycles ledger.
-#[derive(CandidType, Deserialize, Debug, Clone, Eq, PartialEq)]
-pub struct TopUpCyclesLedgerResponse {
-    /// The ledger balance after topping up.
-    ledger_balance: Nat,
-    /// The backend canister cycles after topping up.
-    backend_cycles: Nat,
-    /// The amount topped up.
-    /// - Zero if the ledger balance was already sufficient.
-    /// - The requested amount otherwise.
-    topped_up: Nat,
-}
-
-pub type TopUpCyclesLedgerResult = Result<TopUpCyclesLedgerResponse, TopUpCyclesLedgerError>;
-
 /// Tops up the cycles ledger.
+/// 
+/// # Errors
+/// Errors are enumerated by: `TopUpCyclesLedgerError`
 #[allow(clippy::unused_async)] // TODO: Remove once the code is implemented
 pub async fn top_up_cycles_ledger(request: TopUpCyclesLedgerRequest) -> TopUpCyclesLedgerResult {
     todo!("Add code that tops up the cycles ledger per this request: {request:?}")
